[CI skip] Some refactoring
